requirements.txt
Django==5.1.3
social-auth-app-django==5.4.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1